+++ examples/neo4j/set_classpath.ps1
@@ -1,9 +1,7 @@
-#!/usr/bin/env bash
-
 #############################################################################
 ## 
-##   Set CLASSPATH environment variable for LVM
-##   Last updated on March 14, 2023
+##   Set CLASSPATH environment variable for SWI-Prolog
+##   Last updated on July 21, 2023
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2023 Paulo Moura <pmoura@logtalk.org>
@@ -23,10 +21,11 @@
 #############################################################################
 
 
-CLASSPATH="$(lvmpl --src-root)/../library/jni/dependencies/jpl.jar"
+#Requires -Version 7.3
 
-for jar in jars/*.jar; do
-	CLASSPATH="$PWD/$jar":$CLASSPATH
-done
+Get-ChildItem -Path C:\neo4j-community-5.5.0\lib\* -Filter *.jar |
+Foreach-Object {
+	$classpath += ";" + $_.FullName
+}
 
-export CLASSPATH
+[System.Environment]::setEnvironmentVariable("CLASSPATH", $classpath, "Process")